Use IsUnrecognizedConstraintError (#1680)
This is to decouple Gatekeeper from the error implementation in
frameworks. With this change, frameworks can swap out the implementation
without Gatekeeper caring. Gatekeeper doesn't use any of the fields of
this error type, so it shouldn't rely on the specific type.

Note that this requires upgrading frameworks/ to a version which
provides this method.

Signed-off-by: Will Beason <willbeason@google.com>