Update golangci-lint
The new golangci-lint is more strict about variable naming reqiurements.
It has false negatives on some of our duration variables, so I've
renamed them.

We _could_ suppress the linter in these cases, but it's easier to just
rename the variables.

The other fixes are making all variables in a block of declarations have
explicit types if one does. This seems reasonable since not everyone is
familiar with the convention (which is dependent on the type of
declaration).

Signed-off-by: Will Beason <willbeason@google.com>