Improve error messages for Assertions (#1726)
* Improve error messages for Assertions

Specifically for Assertions which define "Message". Now the test failure
properly notifies the user that the Assertion was looking for violations
for a particular message.

Also add several unit tests for Assertions to cover these cases. Unit
test coverage is now 100% for assertions.go. We could probably add more
tests for assertions, but I'd rather that be its own PR. What I've added
is for the behavior changed by this PR.

Also add validation that "violations" isn't set to a negative number.

Fixes #1639

Signed-off-by: Will Beason <willbeason@google.com>

* Fix linter errors

Signed-off-by: Will Beason <willbeason@google.com>