refactor: Change `gator test` to `gator verify` (#1799)
In this weeks community meeting, we discussed naming that would
differentiate the current `gator test` from a new command that validates
entire manifests against sets of policies (#1786).  We decided to align
gator's commands with the verbs used in conftest.

Accordingly, what has previously been known as `gator test` will be
henceforth called:

`gator verify`

This PR changes package names and the name of the command.  The
aforementioned PR will rebase on this PR to include the name changes.

Signed-off-by: juliankatz <juliankatz@google.com>