Make ConstraintTemplate tests less flaky
We've seen flakiness from these tests for a while, and something in our
changes to frameworks/ seems to be making things more flaky. So I've
included changes that make the ConstraintTemplate tests much less
flaky.

One idea at work is to make tests independent. Several of the tests
implicitly required other tests to pass in order to run properly. I've
removed several of these dependencies so it's more obvious where a
failure actually is. Since these are integration tests, it is easier to
debug when the only failure you see is that a Constraint was not created
properly- a later test failing because a Constraint which wasn't added
is also not enforced is just noise. So in some cases I've merged tests
together and in others I've made each test set up its own environment.

Another is that each call to client.Client methods needs a completely
new object. Reusing objects results in cross-talk between tests.

Signed-off-by: Will Beason <willbeason@google.com>