feat: Add gator test (formerly gator validate) (#1786)
This PR adds a new command to the `gator` CLI:

```
gator test
```

The design doc for this command can be found here:

https://docs.google.com/document/d/1B0hXDia8SExOkCVAEbPVHFJmnWeJgHziGvMxnzrENa0

This command ingests k8s resources, templates, and constraints then produces
any violations found.  This can be used in a "shift-left" CI flow to present
users with policy failures earlier in their development cycle.

In development, this command was known as `gator validate`.  It replaced the existing `gator test`, which was renamed to `gator verify`.