Reference `gator verify` in `gator test --help`
While we are at liberty to change the `gator` CLI and its subcommands
at will during the "alpha" phase, we have nonetheless created a
confusing situation:

We've replaced `gator test` with a different command, and re-named the
original `gator test` to `gator verify`.

This PR adds some information in the (new) `gator test`'s "help"
message, pointing users to the new command.

Fix #1835

Signed-off-by: juliankatz <juliankatz@google.com>