Return violating resource in pkg/gator/test.Test
During the refactoring of the Constraint Framework, the violating
resource was removed from the pkg/types.Result struct.

As this struct was returned in the public interface of Gatekeeper's
pkg/gator/test.Test() function, consumers of that test package were
deprived of their ability to discern which resource was causing the
violation returned to them.

This PR declares a new set of types in the `test` package that are meant
to mirror those of CF's `types` package, but that include a reference to
the violating object.

This PR also updates the gator test CLI command to use this new
interface.

Signed-off-by: juliankatz <juliankatz@google.com>