feat: Allow disabling of referential data in gatortest.Test()
In manual benchmarks of the gator test CLI, I found that calls to Test()
that included adding all the objects under evaluation to OPA via the
AddData() call were exponentially slower.

These calls to AddData() are to support referential templates.

This PR adds a variadic argument to Test() that allows one to disable
referential data support in Test(), yielding a significant speed
improvement for users who do not require the feature.

Next steps:

1) Add a flag to `gator test` like --referential-data, that defaults to
   true

Signed-off-by: juliankatz <juliankatz@google.com>