docs: add a dedicated page in the website for Kubernetes compatibility.
Summary:

For improved discovery, this moves the matrix from the release management
file in the Gatekeeper repo, to the Gatekeeper website under the Architecture
heading. For better readability and simpler maintenance, instead of using a
matrix, we are using a 2 column table. This way, the team won't need to add a
new row and column to the matrix for each new release.

Reviewers: ritazh

Signed-off-by: KayenM <kayenkmehta@gmail.com>